requirements.txt
streamlit
numpy
pandas
pyyaml
sentence-transformers
faiss-cpu
rank-bm25
pymupdf
tqdm